+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+Unreleased
+- Retry throttling errors on XML/Query services (SES, SQS, SNS)
+
 v2.7.0 (2026-03-04)
 - Upgrade hackney to 3.x
 - Correctly handle pod identity authorization when using Req adapter